+++ connection/connection.cpp
@@ -198,7 +198,7 @@ void QubicConnection::receiveAFullPacket(RequestResponseHeader& header, std::vec
 void QubicConnection::sendEndPacket(uint32_t dejavu)
 {
     RequestResponseHeader nop{};
-    nop.setType(35);
+    nop.setType(END_RESPONSE);
     if (dejavu != 0xffffffff) nop.setDejavu(dejavu);
     else nop.randomizeDejavu();
     nop.setSize(sizeof(RequestResponseHeader));
@@ -258,6 +258,10 @@ int QubicConnection::enqueueSend(uint8_t* buffer, int sz)
         if (dejavu)
         {
             requestMapperFrom.add(dejavu, buffer, sz, nullptr);
+            if (isLogRequestType(header.type()))
+            {
+                incLogReqSent();
+            }
         }
     }
     mBuffer->EnqueuePacket(buffer);